Add files via upload
Step 4:tune layers and neurons.To be specific,I tuned 1 layer with neurons[2,3,4,5,7,10,15,20,25,30,40,50,70,100], 2 layer with neurons[[2,2],[3,3],[4,4],[5,5],[10,10],[30,30],[50,50],[70,70],[100,100],[2,4],[4,8],[8,16],[16,32],[32,64]], 3 layer with neurons[[2,2,2],[3,3,3],[4,4,4],[5,5,5],[10,10,10],[30,30,30],[50,50,50],[70,70,70]].And 1 layer with 7 neurons wins.  
PS: as this note book is over 25 mb which is the limit of uplaoding size for github,I delete to graphs the results in it,please refer to excel file"Result of Tuning" for tuning results.